+++ R/add_noise.R
@@ -10,6 +10,7 @@
 #' @param max.iterations Numeric vector of length 1. Specifies the maximum number of iterations that the internal signal-to-noise adjusting routine will run before stopping. Note that in most cases the default maximum number of iterations (1000) is not reached.
 #' @param kind Character vector of length 1 indicating the kind of noise, “white”, “pink”, “power”, "brown", or “red”. Noise is synthesized with a modified version of the function \code{\link[tuneR]{noise}}. Default is "pink" which is similar to background noise in natural environments.
 #' @param alpha Numeric vector of length 1. The power for the power law noise (defaults are 1 for pink and 1.5 for red noise). Only used when \code{kind = "power"}.
+#' @param seed Numeric vector of length 1. Seed for random number generation. Default is 123. If NULL, the seed is not set.
 #' @param ... Additional arguments to be passed internally to \code{\link{signal_to_noise_ratio}}.
 #' @return Object 'X' in which the wave objects have been modified to match the target signal-to-noise ratio. It also includes an additional column, 'adjusted.snr', with the new signal-to-noise ratio values.
 #' @export
@@ -43,6 +44,7 @@ add_noise <-
            max.iterations = 1000,
            kind = c("pink", "white", "brown", "red", "power"),
            alpha = 1,
+           seed = 123,
            ...) {
     
     # assign a value to kind
@@ -79,7 +81,8 @@ add_noise <-
         max.iterations = max.iterations,
         Y = X,
         kind = kind,
-        alpha = alpha
+        alpha = alpha, 
+        seed = seed
       )
     
     names(wav_snr_list) <- target_rows