+++ NEWS.md
@@ -8,6 +8,7 @@ baRulho 2.1.5
 
 ### MINOR IMPROVEMENTS
 
+* Improve error messages for functions working with master annotations
 * Fix bug in `add_noise()` in which '...' were not sent internally to `signal_to_noise_ratio()`
 * Improved error message when missing columns in supplied `Y` data frames